refactor: add setup.py for pip install -e
