+++ README.md
@@ -1,47 +1,93 @@
 # ROS Map Editor
 
+A GUI tool for editing ROS map files in PGM format. This tool allows you to view and modify occupancy grid maps used in ROS (Robot Operating System) navigation.
 
+## Acknowledgements
 
-A Python Qt based tool for editing maps generated by [ROS gmapping package](http://wiki.ros.org/gmapping).
+- [https://github.com/TheOnceAndFutureSmalltalker/ros_map_editor](https://github.com/TheOnceAndFutureSmalltalker/ros_map_editor)
 
-![ROS Map Editor](https://github.com/TheOnceAndFutureSmalltalker/ros_map_editor/blob/master/images/ui.PNG)
+![ROS Map Editor](images/usage.gif)
 
-This was inspired by lack of any editors dedicated to this purpose.  All chatter I could find just mentioned using Gimp, or some equivalent for editing these files, which is quite inadequate.  
+## Features
 
-In particular, I needed a way to add areas of map which are "out of bounds" for the robot, but not necessarily obstacles detected by sensor.  Things like ponds, sand traps, flower beds, etc.  Flat surfaces, not detected by sensors in mapping, but off limits (or highly dangerous) to the robot.  
+- Load and display PGM map files used in ROS
+- Edit map cells (occupied, free, unknown)
+- Zoom in/out for detailed editing
+- Minimap for navigation in large maps
+- Automatically saves changes on exit
 
-I also needed a way to match the map with a real world map like Google Maps and associate grid cells with lat/lng.
+## Installation
 
-Also, just general cleanup of map is an arduous taks with Gimp, et al.
+### From PyPI
 
-With the Map Editor tool, I can now do all of these things (well, some of these features have yet to be added).
+```bash
+pip install ros-map-editor
+```
 
-## Current Functionality
+### From Source
 
-1) Edits files of type pgm with an accompanying yaml file describing metadata as generated by ROS gmapping package
+```bash
+git clone https://github.com/tianbot/ros-map-editor.git
+cd ros-map-editor
+pip install -e .
+```
 
-2) Grid cells can be set to occupied (black), unoccupied (white), or uncertain (gray).
+## Usage
 
-3) Thumb image shows entire map with the current viewport indicated (weirdly, commercial packages did not have this)
+### Command Line
 
-4) Can zoom 100%, 200%, 400%, 800%, and 1600% for large grain to fine grain work
+```bash
+ros-map-editor path/to/your/map
+```
 
-5) Set color choice to white (unoccupied), black (occupied), gray (uncertain), and alternating (white->black->gray).
+or simply:
 
-6) Can color free form with mouse drag or pixel by pixel by clicking
+```bash
+python3 -m ros_map_editor path/to/your/map
+```
 
-## Future Functionality
+The tool will automatically look for the corresponding YAML file (map.yaml) in the same directory.
 
-1) Draw rectangles
+### Interface
 
-2) Overlay Google Map
+- **Zoom**: Select zoom level from the dropdown menu
+- **Color**: Choose what to draw:
+  - **Occupied**: Mark cells as obstacles (black)
+  - **Free**: Mark cells as free space (white)
+  - **Unknown**: Mark cells as unknown (gray)
+  - **Alternate**: Cycle through states with each click
+- **Minimap**: Shows your current position in the map with a red rectangle
+- **Drawing**: Click or drag to modify cells
 
-3) Start a New map (currently only opens on current map)
+## Map File Format
 
-4) Associate map/grids with lat/lng
+The tool works with standard ROS map files:
 
-5) mods to yaml file as well
+- A PGM image file containing the occupancy grid
+- A YAML file with metadata (resolution, origin, thresholds)
 
+Example YAML file:
 
+```yaml
+image: map.pgm
+resolution: 0.050000
+origin: [-10.000000, -10.000000, 0.000000]
+negate: 0
+occupied_thresh: 0.65
+free_thresh: 0.196
+```
 
-## Getting Started
+## Requirements
+
+- Python 3.6+
+- PyQt5
+- Pillow (PIL)
+- PyYAML
+
+## License
+
+Public domain
+
+## Contributing
+
+Contributions are welcome! Please feel free to submit a Pull Request.
