+++ requirements.txt
@@ -1,0 +1,3 @@
+PyQt5>=5.15.0
+Pillow>=8.0.0
+PyYAML>=5.1.0
