+++ ros_map_editor/__init__.py
@@ -1,0 +1,6 @@
+"""
+ROS Map Editor package
+A GUI tool for editing ROS map files in PGM format
+"""
+
+__version__ = "0.1.0"
