--- ui_map_editor.py
@@ -1,225 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'UI_MapEditor_2.ui'
-#
-# Created by: PyQt5 UI code generator 5.11.3
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_MapEditor(object):
-    def setupUi(self, MapEditor):
-        MapEditor.setObjectName("MapEditor")
-        MapEditor.resize(939, 747)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(MapEditor.sizePolicy().hasHeightForWidth())
-        MapEditor.setSizePolicy(sizePolicy)
-        self.centralwidget = QtWidgets.QWidget(MapEditor)
-        self.centralwidget.setObjectName("centralwidget")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.centralwidget)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.graphicsView = QtWidgets.QGraphicsView(self.centralwidget)
-        self.graphicsView.setObjectName("graphicsView")
-        self.verticalLayout.addWidget(self.graphicsView)
-        self.horizontalLayout_bottom = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_bottom.setObjectName("horizontalLayout_bottom")
-
-        self.image_preview_group = QtWidgets.QGroupBox(self.centralwidget)
-        self.image_preview_group.setObjectName("image_preview_group")
-        self.preview_layout = QtWidgets.QHBoxLayout(self.image_preview_group)
-        self.preview_layout.setObjectName("preview_layout")
-        self.label_2 = QtWidgets.QLabel(self.image_preview_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_2.sizePolicy().hasHeightForWidth())
-        self.label_2.setSizePolicy(sizePolicy)
-        self.label_2.setMaximumSize(QtCore.QSize(150, 150))
-        self.label_2.setBaseSize(QtCore.QSize(100, 100))
-        self.label_2.setFrameShape(QtWidgets.QFrame.Box)
-        self.label_2.setText("")
-        self.label_2.setScaledContents(True)
-        self.label_2.setObjectName("label_2")
-        self.preview_layout.addWidget(self.label_2)
-        self.horizontalLayout_bottom.addWidget(self.image_preview_group)
-
-        self.info_group = QtWidgets.QGroupBox(self.centralwidget)
-        self.info_group.setObjectName("info_group")
-        self.formLayout = QtWidgets.QFormLayout(self.info_group)
-        self.formLayout.setFieldGrowthPolicy(QtWidgets.QFormLayout.ExpandingFieldsGrow)
-        self.formLayout.setObjectName("formLayout")
-        self.label_3 = QtWidgets.QLabel(self.info_group)
-        self.label_3.setObjectName("label_3")
-        self.formLayout.setWidget(0, QtWidgets.QFormLayout.LabelRole, self.label_3)
-        self.filename_lbl = QtWidgets.QLabel(self.info_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.filename_lbl.sizePolicy().hasHeightForWidth())
-        self.filename_lbl.setSizePolicy(sizePolicy)
-        self.filename_lbl.setObjectName("filename_lbl")
-        self.formLayout.setWidget(0, QtWidgets.QFormLayout.FieldRole, self.filename_lbl)
-        self.label_4 = QtWidgets.QLabel(self.info_group)
-        self.label_4.setObjectName("label_4")
-        self.formLayout.setWidget(1, QtWidgets.QFormLayout.LabelRole, self.label_4)
-        self.width_lbl = QtWidgets.QLabel(self.info_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.width_lbl.sizePolicy().hasHeightForWidth())
-        self.width_lbl.setSizePolicy(sizePolicy)
-        self.width_lbl.setObjectName("width_lbl")
-        self.formLayout.setWidget(1, QtWidgets.QFormLayout.FieldRole, self.width_lbl)
-        self.label_5 = QtWidgets.QLabel(self.info_group)
-        self.label_5.setObjectName("label_5")
-        self.formLayout.setWidget(2, QtWidgets.QFormLayout.LabelRole, self.label_5)
-        self.height_lbl = QtWidgets.QLabel(self.info_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.height_lbl.sizePolicy().hasHeightForWidth())
-        self.height_lbl.setSizePolicy(sizePolicy)
-        self.height_lbl.setObjectName("height_lbl")
-        self.formLayout.setWidget(2, QtWidgets.QFormLayout.FieldRole, self.height_lbl)
-        self.label_9 = QtWidgets.QLabel(self.info_group)
-        self.label_9.setObjectName("label_9")
-        self.formLayout.setWidget(3, QtWidgets.QFormLayout.LabelRole, self.label_9)
-        self.cell_lbl = QtWidgets.QLabel(self.info_group)
-        self.cell_lbl.setObjectName("cell_lbl")
-        self.formLayout.setWidget(3, QtWidgets.QFormLayout.FieldRole, self.cell_lbl)
-        self.horizontalLayout_bottom.addWidget(self.info_group)
-
-        self.controls_group = QtWidgets.QGroupBox(self.centralwidget)
-        self.controls_group.setObjectName("controls_group")
-        self.controls_layout_vertical = QtWidgets.QVBoxLayout(self.controls_group)
-        self.controls_layout_vertical.setObjectName("controls_layout_vertical")
-
-        self.controls_layout_row1 = QtWidgets.QHBoxLayout()
-        self.controls_layout_row1.setObjectName("controls_layout_row1")
-
-        self.color_label = QtWidgets.QLabel(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.color_label.sizePolicy().hasHeightForWidth())
-        self.color_label.setSizePolicy(sizePolicy)
-        self.color_label.setObjectName("color_label")
-        self.controls_layout_row1.addWidget(self.color_label)
-
-        self.colorBox = QtWidgets.QComboBox(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.colorBox.sizePolicy().hasHeightForWidth())
-        self.colorBox.setSizePolicy(sizePolicy)
-        self.colorBox.setObjectName("colorBox")
-        self.controls_layout_row1.addWidget(self.colorBox)
-
-        self.zoom_label = QtWidgets.QLabel(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.zoom_label.sizePolicy().hasHeightForWidth())
-        self.zoom_label.setSizePolicy(sizePolicy)
-        self.zoom_label.setObjectName("zoom_label")
-        self.controls_layout_row1.addWidget(self.zoom_label)
-
-        self.zoomBox = QtWidgets.QComboBox(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.zoomBox.sizePolicy().hasHeightForWidth())
-        self.zoomBox.setSizePolicy(sizePolicy)
-        self.zoomBox.setObjectName("zoomBox")
-        self.controls_layout_row1.addWidget(self.zoomBox)
-
-        self.controls_layout_vertical.addLayout(self.controls_layout_row1)
-
-        self.controls_layout_row2 = QtWidgets.QHBoxLayout()
-        self.controls_layout_row2.setObjectName("controls_layout_row2")
-
-        self.focusButton = QtWidgets.QPushButton(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.focusButton.sizePolicy().hasHeightForWidth())
-        self.focusButton.setSizePolicy(sizePolicy)
-        self.focusButton.setObjectName("focusButton")
-        self.controls_layout_row2.addWidget(self.focusButton)
-
-        self.saveButton = QtWidgets.QPushButton(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.saveButton.sizePolicy().hasHeightForWidth())
-        self.saveButton.setSizePolicy(sizePolicy)
-        self.saveButton.setObjectName("saveButton")
-        self.controls_layout_row2.addWidget(self.saveButton)
-
-        self.closeButton = QtWidgets.QPushButton(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.closeButton.sizePolicy().hasHeightForWidth())
-        self.closeButton.setSizePolicy(sizePolicy)
-        self.closeButton.setObjectName("closeButton")
-        self.controls_layout_row2.addWidget(self.closeButton)
-
-        self.boxSelectCheck = QtWidgets.QCheckBox(self.controls_group)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.boxSelectCheck.sizePolicy().hasHeightForWidth())
-        self.boxSelectCheck.setSizePolicy(sizePolicy)
-        self.boxSelectCheck.setObjectName("mode_select")
-        self.controls_layout_row2.addWidget(self.boxSelectCheck)
-
-        self.controls_layout_vertical.addLayout(self.controls_layout_row2)
-
-        self.horizontalLayout_bottom.addWidget(self.controls_group)
-        self.verticalLayout.addLayout(self.horizontalLayout_bottom)
-
-        MapEditor.setCentralWidget(self.centralwidget)
-        self.menubar = QtWidgets.QMenuBar(MapEditor)
-        self.menubar.setGeometry(QtCore.QRect(0, 0, 939, 21))
-        self.menubar.setObjectName("menubar")
-        MapEditor.setMenuBar(self.menubar)
-        self.statusbar = QtWidgets.QStatusBar(MapEditor)
-        self.statusbar.setObjectName("statusbar")
-        MapEditor.setStatusBar(self.statusbar)
-
-        self.retranslateUi(MapEditor)
-        QtCore.QMetaObject.connectSlotsByName(MapEditor)
-
-    def retranslateUi(self, MapEditor):
-        _translate = QtCore.QCoreApplication.translate
-        MapEditor.setWindowTitle(_translate("MapEditor", "MainWindow"))
-        self.image_preview_group.setTitle(_translate("MapEditor", "Preview"))
-        self.info_group.setTitle(_translate("MapEditor", "Image Information"))
-        self.label_3.setText(_translate("MapEditor", "File"))
-        self.filename_lbl.setText(_translate("MapEditor", "TextLabel"))
-        self.label_4.setText(_translate("MapEditor", "Width"))
-        self.width_lbl.setText(_translate("MapEditor", "TextLabel"))
-        self.label_5.setText(_translate("MapEditor", "Height"))
-        self.height_lbl.setText(_translate("MapEditor", "TextLabel"))
-        self.label_9.setText(_translate("MapEditor", "Cell"))
-        self.cell_lbl.setText(_translate("MapEditor", "TextLabel"))
-        self.controls_group.setTitle(_translate("MapEditor", "Controls"))
-        self.color_label.setText(_translate("MapEditor", "Color"))
-        self.zoom_label.setText(_translate("MapEditor", "Zoom"))
-        self.focusButton.setText(_translate("MapEditor", "Center"))
-        self.saveButton.setText(_translate("MapEditor", "Save"))
-        self.closeButton.setText(_translate("MapEditor", "Close"))
-        self.boxSelectCheck.setText(_translate("MapEditor", "Selection Mode"))
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    MapEditor = QtWidgets.QMainWindow()
-    ui = Ui_MapEditor()
-    ui.setupUi(MapEditor)
-    MapEditor.show()
-    sys.exit(app.exec_())
